fix: satisfy strict mode build
